[BEAM-5495] Move PipelineResources to org.apache.beam.runners.core.construction.resources package
