+++ runners/portability/java/src/main/java/org/apache/beam/runners/portability/PortableRunner.java
@@ -17,7 +17,7 @@
  */
 package org.apache.beam.runners.portability;
 
-import static org.apache.beam.runners.core.construction.PipelineResources.detectClassPathResourcesToStage;
+import static org.apache.beam.runners.core.construction.resources.PipelineResources.detectClassPathResourcesToStage;
 import static org.apache.beam.vendor.guava.v26_0_jre.com.google.common.base.Preconditions.checkState;
 
 import java.io.File;
